Revert "Revert "fix: (12)-refactored global style sheet (#6436)" (#6491)"
This reverts commit 73254952d9e3ae753d4c39e35ff407e1bebba082.
