--- docs/docs/auto-docs/PledgeModal/functions/getMemberLabel.md
@@ -1,33 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Function: getMemberLabel()
-
-> **getMemberLabel**(`member`): `string`
-
-Defined in: [src/screens/UserPortal/Campaigns/PledgeModal.tsx:104](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/UserPortal/Campaigns/PledgeModal.tsx#L104)
-
-Builds a display label for a member.
-Empty name parts are safely ignored.
-
-## Parameters
-
-### member
-
-[`InterfaceUserInfoPG`](../../utils/interfaces/interfaces/InterfaceUserInfoPG.md)
-
-User object containing name fields
-
-## Returns
-
-`string`
-
-Full name string constructed from available name parts
-
-## Example
-
-```ts
-getMemberLabel({ firstName: 'John', lastName: 'Doe' } as InterfaceUserInfoPG);
-// returns "John Doe"
-```
