--- docs/docs/auto-docs/PledgeModal/interfaces/InterfacePledgeModal-1.md
@@ -1,71 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Interface: InterfacePledgeModal
-
-Defined in: [src/screens/UserPortal/Campaigns/PledgeModal.tsx:67](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/UserPortal/Campaigns/PledgeModal.tsx#L67)
-
-## Properties
-
-### campaignId
-
-> **campaignId**: `string`
-
-Defined in: [src/screens/UserPortal/Campaigns/PledgeModal.tsx:70](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/UserPortal/Campaigns/PledgeModal.tsx#L70)
-
-***
-
-### hide()
-
-> **hide**: () => `void`
-
-Defined in: [src/screens/UserPortal/Campaigns/PledgeModal.tsx:69](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/UserPortal/Campaigns/PledgeModal.tsx#L69)
-
-#### Returns
-
-`void`
-
-***
-
-### isOpen
-
-> **isOpen**: `boolean`
-
-Defined in: [src/screens/UserPortal/Campaigns/PledgeModal.tsx:68](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/UserPortal/Campaigns/PledgeModal.tsx#L68)
-
-***
-
-### mode
-
-> **mode**: `"create"` \| `"edit"`
-
-Defined in: [src/screens/UserPortal/Campaigns/PledgeModal.tsx:74](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/UserPortal/Campaigns/PledgeModal.tsx#L74)
-
-***
-
-### pledge
-
-> **pledge**: [`InterfacePledgeInfo`](../../utils/interfaces/interfaces/InterfacePledgeInfo.md)
-
-Defined in: [src/screens/UserPortal/Campaigns/PledgeModal.tsx:72](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/UserPortal/Campaigns/PledgeModal.tsx#L72)
-
-***
-
-### refetchPledge()
-
-> **refetchPledge**: () => `void`
-
-Defined in: [src/screens/UserPortal/Campaigns/PledgeModal.tsx:73](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/UserPortal/Campaigns/PledgeModal.tsx#L73)
-
-#### Returns
-
-`void`
-
-***
-
-### userId
-
-> **userId**: `string`
-
-Defined in: [src/screens/UserPortal/Campaigns/PledgeModal.tsx:71](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/UserPortal/Campaigns/PledgeModal.tsx#L71)
