--- docs/docs/auto-docs/PledgeModal/variables/default-1.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: default
-
-> `const` **default**: `React.FC`\<[`InterfacePledgeModal`](../interfaces/InterfacePledgeModal-1.md)\>
-
-Defined in: [src/screens/UserPortal/Campaigns/PledgeModal.tsx:107](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/UserPortal/Campaigns/PledgeModal.tsx#L107)
