+++ docs/docs/auto-docs/screens/UserPortal/UserScreen/UserScreen/functions/default.md
@@ -6,7 +6,7 @@
 
 > **default**(): `Element`
 
-Defined in: [src/screens/UserPortal/UserScreen/UserScreen.tsx:68](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/UserPortal/UserScreen/UserScreen.tsx#L68)
+Defined in: [src/screens/UserPortal/UserScreen/UserScreen.tsx:67](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/UserPortal/UserScreen/UserScreen.tsx#L67)
 
 ## Returns
