+++ public/locales/hi/translation.json
@@ -922,7 +922,8 @@
     "pledgeDate": "प्रतिज्ञा तिथि",
     "pledged": "प्रतिबद्ध",
     "donated": "दान किया गया",
-    "campaignNotActive": "अभियान वर्तमान में सक्रिय नहीं है"
+    "campaignNotActive": "अभियान वर्तमान में सक्रिय नहीं है",
+    "close": "बंद करें"
   },
   "createPostModal": {
     "title": "पदों",
