+++ public/locales/zh/translation.json
@@ -905,7 +905,8 @@
     "pledgeDate": "承诺日期",
     "pledged": "已承诺",
     "donated": "已捐赠",
-    "campaignNotActive": "活动当前未激活"
+    "campaignNotActive": "活动当前未激活",
+    "close": "关闭"
   },
   "createPostModal": {
     "title": "帖子",
