+++ src/screens/UserPortal/Campaigns/Campaigns.tsx
@@ -7,7 +7,7 @@
  */
 import React, { useCallback, useMemo, useState } from 'react';
 import { Button } from 'react-bootstrap';
-import styles from '../../../style/app-fixed.module.css';
+import styles from './Campaigns.module.css';
 import { useTranslation } from 'react-i18next';
 import { Navigate, useNavigate, useParams } from 'react-router';
 import { Campaign, WarningAmberRounded } from '@mui/icons-material';
