+++ src/screens/UserPortal/Campaigns/PledgeModal.module.css
@@ -1,0 +1,53 @@
+.pledgeModal {
+  max-width: 80vw;
+  margin-top: 2vh;
+  margin-left: 13vw;
+}
+
+.titleModal {
+  color: var(--color-gray-500);
+  font-weight: var(--font-weight-semibold);
+  font-size: var(--font-size-3xl);
+  width: 65%;
+  margin-bottom: 0;
+}
+
+.modalCloseBtn {
+  width: var(--space-9);
+  height: var(--space-9);
+  padding: var(--space-5);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.noOutline input,
+.noOutline:is(:hover, :focus, :active, :focus-visible, .show) {
+  outline: none;
+}
+
+.noOutline input:disabled {
+  -webkit-text-fill-color: var(--color-gray-600) !important;
+}
+
+.noOutline textarea:disabled {
+  color: var(--color-gray-600) !important;
+}
+
+.addButton {
+  margin-bottom: var(--space-4);
+  color: var(--color-gray-700);
+  background-color: var(--color-blue-200);
+  border-color: var(--color-blue-200);
+  --bs-btn-focus-box-shadow: none;
+}
+
+.addButton:is(:hover, :active, :focus) {
+  background-color: var(--color-blue-500);
+  border-color: var(--color-gray-700);
+}
+
+.addButton:disabled {
+  background-color: var(--color-gray-50);
+  border-color: var(--color-blue-200);
+}
