+++ src/screens/UserPortal/UserScreen/UserScreen.module.css
@@ -1,175 +1,45 @@
-.pageContainer {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-  padding: var(--space-5) var(--space-7) var(--space-0)
-    calc(var(--space-20) + var(--space-8) + var(--space-7));
+@keyframes moveLeft {
+  from {
+    transform: translateX(var(--space-11));
+  }
+
+  to {
+    transform: translateX(0);
+  }
+}
+
+@keyframes moveRight {
+  from {
+    transform: translateX(calc(-1 * var(--space-11)));
+  }
+
+  to {
+    transform: translateX(0);
+  }
 }
 
 .expand {
+  margin-left: var(--space-13);
   padding-left: var(--space-11);
   animation: moveLeft 0.9s ease-in-out;
 }
-.avatarStyle {
-  border-radius: 100%;
-}
-.profileContainer {
-  border: none;
-  padding: var(--space-8) var(--space-3);
-  height: var(--space-10);
-  border-radius: 8px 0px 0px 8px;
-  display: flex;
-  align-items: center;
-  background-color: white !important;
-  box-shadow:
-    0 4px 4px 0 rgba(177, 177, 177, 0.2),
-    0 6px 20px 0 rgba(151, 151, 151, 0.19);
-}
-.profileContainer:focus {
-  outline: none;
-  background-color: var(--bs-gray-100);
-}
-.imageContainer {
-  width: var(--space-11);
-}
-.profileContainer .profileText {
-  flex: 1;
-  text-align: start;
-  overflow: hidden;
-  margin-right: var(--space-2);
-}
-.angleDown {
-  margin-left: var(--space-2);
-}
-.profileContainer .profileText .primaryText {
-  font-size: var(--font-size-md);
-  font-weight: var(--font-weight-semibold);
-  overflow: hidden;
-  display: -webkit-box;
-  -webkit-line-clamp: 2; /* number of lines to show */
-  -webkit-box-orient: vertical;
-  word-wrap: break-word;
-  white-space: normal;
-}
-.profileContainer .profileText .secondaryText {
-  font-size: var(--font-size-xs);
-  font-weight: var(--font-weight-regular);
-  color: var(--bs-secondary);
-  display: block;
-  text-transform: capitalize;
-}
 
 .contract {
-  padding-left: calc(var(--space-20) + var(--space-8) + var(--space-7));
+  padding-left: var(--space-22);
   animation: moveRight 0.5s ease-in-out;
 }
 
-.collapseSidebarButton {
-  position: fixed;
-  height: var(--space-9);
-  bottom: 0;
-  z-index: 9999;
-  width: calc(var(--space-20) + var(--space-8));
-  background-color: rgba(245, 245, 245, 0.7);
-  color: black;
-  border: none;
-  border-radius: 0;
-}
-
-.collapseSidebarButton:hover,
-.opendrawer:hover {
-  opacity: 1;
-  color: black !important;
-}
-.opendrawer {
-  position: fixed;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  top: 0;
-  left: 0;
-  width: var(--space-9);
-  height: 100vh;
-  z-index: 9999;
-  background-color: rgba(245, 245, 245);
-  border: none;
-  border-radius: 0;
-  margin-right: var(--space-6);
-  color: black;
-}
-.profileDropdown {
-  background-color: transparent !important;
-}
-.profileDropdown .dropdown-toggle .btn .btn-normal {
-  display: none !important;
-  background-color: transparent !important;
-}
-.dropdownToggle {
-  background-image: url(/public/images/svg/angleDown.svg);
-  background-repeat: no-repeat;
-  background-position: center;
-  background-color: azure;
-}
-
-.dropdownToggle::after {
-  border-top: none !important;
-  border-bottom: none !important;
-}
-
-.opendrawer:hover {
-  transition: background-color 0.5s ease;
-  background-color: var(--bs-primary);
-}
-.collapseSidebarButton:hover {
-  transition: background-color 0.5s ease;
-  background-color: var(--bs-primary);
-}
-
-@media (max-width: var(--breakpoint-1120)) {
+@media (max-width: 1120px) {
   .contract {
-    padding-left: calc(var(--space-19) + var(--space-8) + var(--space-7));
-  }
-  .collapseSidebarButton {
-    width: calc(var(--space-18) + var(--space-8));
-  }
-}
-
-@media (max-height: var(--breakpoint-900)) {
-  .collapseSidebarButton {
-    width: calc(var(--space-20) + var(--space-5));
-  }
-}
-@media (max-height: var(--breakpoint-650)) {
-  .pageContainer {
-    padding: var(--space-5) var(--space-7) var(--space-0) var(--space-19);
-  }
-  .collapseSidebarButton {
-    width: var(--space-18);
-  }
-  .opendrawer {
-    width: var(--space-8);
+    padding-left: var(--space-21);
   }
 }
 
 /* For tablets */
-@media (max-width: var(--breakpoint-820)) {
-  .pageContainer {
-    padding-left: var(--space-9);
-  }
-
-  .opendrawer {
-    width: var(--space-7);
-  }
-
+@media (max-width: 820px) {
   .contract,
   .expand {
     animation: none;
-  }
-
-  .collapseSidebarButton {
-    width: 100%;
-    left: 0;
-    right: 0;
   }
 }
 
@@ -178,5 +48,19 @@
 }
 
 .titleContainer {
-  flex: 1;
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-4);
+}
+
+.titleContainer h1 {
+  font-size: var(--font-size-28);
+  font-weight: var(--font-weight-bold);
+  color: var(--color-gray-500);
+  margin-top: var(--space-2);
+}
+
+.drawer {
+  position: relative;
+  height: 100vh;
 }
